# frontend/apps/site/openapi-spec.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "CloudCrafter API",
		"version": "v1"
	},
	"paths": {
...
		"/api/Stacks": {
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostCreateStack",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/provider": {
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostCreateStackFromSourceProvider",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackFromSourceProviderCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackFromSourceProviderCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackFromSourceProviderCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/StackCreatedDto"
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/{id}": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetStackDetail",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Stacks"],
				"operationId": "DeleteStack",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			},
			"put": {
				"tags": ["Stacks"],
				"operationId": "UpdateStack",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/StackDetailDto"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/{id}/deploy": {
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostDispatchStackDeployment",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentCreatedDetailsDto"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentCreatedDetailsDto"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/DeploymentCreatedDetailsDto"
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/{id}/deployments": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetDeploymentsForStack",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleDeploymentDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleDeploymentDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/SimpleDeploymentDto"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/deployments/{deploymentId}/logs": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetDeploymentLogs",
				"parameters": [
					{
						"name": "deploymentId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DeploymentLogDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DeploymentLogDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DeploymentLogDto"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/Stacks/{stackId}/services/{stackServiceId}": {
			"patch": {
				"tags": ["Stacks"],
				"operationId": "UpdateStackService",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "stackServiceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/environment-variables": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetEnvironmentVariables",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "includeSecrets",
						"in": "query",
						"schema": {
							"type": "boolean",
							"default": false
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableDto"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostCreateEnvironmentVariable",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Created"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/environment-variables/{id}": {
			"put": {
				"tags": ["Stacks"],
				"operationId": "PutUpdateEnvironmentVariable",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Stacks"],
				"operationId": "DeleteEnvironmentVariable",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/environment-variable-groups": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetEnvironmentVariableGroups",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableGroupDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableGroupDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackEnvironmentVariableGroupDto"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostCreateEnvironmentVariableGroup",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableGroupCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableGroupCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackEnvironmentVariableGroupCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Created"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/environment-variable-groups/{id}": {
			"put": {
				"tags": ["Stacks"],
				"operationId": "PutUpdateEnvironmentVariableGroup",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableGroupCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableGroupCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackEnvironmentVariableGroupCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Stacks"],
				"operationId": "DeleteEnvironmentVariableGroup",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/services/{stackServiceId}/volumes": {
			"get": {
				"tags": ["Stacks"],
				"operationId": "GetStackServiceVolumes",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "stackServiceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackServiceVolumeDto"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackServiceVolumeDto"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/StackServiceVolumeDto"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Stacks"],
				"operationId": "PostCreateStackServiceVolume",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "stackServiceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackServiceVolumeCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackServiceVolumeCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStackServiceVolumeCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Created"
					}
				}
			}
		},
		"/api/Stacks/{stackId}/services/{stackServiceId}/volumes/{id}": {
			"put": {
				"tags": ["Stacks"],
				"operationId": "PutUpdateStackServiceVolume",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "stackServiceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceVolumeCommand"
							}
						},
						"text/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceVolumeCommand"
							}
						},
						"application/*+json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStackServiceVolumeCommand"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Stacks"],
				"operationId": "DeleteStackServiceVolume",
				"parameters": [
					{
						"name": "stackId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "stackServiceId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateStackCommand": {
				"required": [
					"name",
					"gitRepository",
					"gitBranch",
					"pathInGitRepository",
					"buildOption",
					"environmentId",
					"serverId"
				],
				"type": "object",
				"properties": {
					"name": {
						"minLength": 3,
						"type": "string"
					},
					"gitRepository": {
						"minLength": 1,
						"type": "string"
					},
					"gitBranch": {
						"minLength": 1,
						"type": "string"
					},
					"pathInGitRepository": {
						"type": "string"
					},
					"buildOption": {
						"$ref": "#/components/schemas/CreateStackBuildOption"
					},
					"environmentId": {
						"type": "string",
						"format": "uuid"
					},
					"serverId": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
			"CreateStackEnvironmentVariableCommand": {
				"required": ["key", "value", "type", "isSecret"],
				"type": "object",
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EnvironmentVariableType"
					},
					"isSecret": {
						"type": "boolean"
					},
					"groupId": {
						"type": "string",
						"format": "uuid",
						"nullable": true
					}
				}
			},
...
			"CreateStackFromSourceProviderCommand": {
				"required": [
					"name",
					"providerId",
					"repositoryId",
					"branch",
					"path",
					"repository",
					"buildOption",
					"environmentId",
					"serverId"
				],
				"type": "object",
				"properties": {
					"name": {
						"minLength": 3,
						"type": "string"
					},
					"providerId": {
						"type": "string",
						"format": "uuid"
					},
					"repositoryId": {
						"minLength": 1,
						"type": "string"
					},
					"branch": {
						"minLength": 1,
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"repository": {
						"type": "string"
					},
					"buildOption": {
						"$ref": "#/components/schemas/CreateStackBuildOption"
					},
					"environmentId": {
						"type": "string",
						"format": "uuid"
					},
					"serverId": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
...
			"DeploymentCreatedDetailsDto": {
				"required": ["deploymentId"],
				"type": "object",
				"properties": {
					"deploymentId": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
			"DeploymentLogDto": {
				"required": ["message", "level", "at"],
				"type": "object",
				"properties": {
					"message": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/ChannelOutputLogLineLevel"
					},
					"at": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"ProblemDetails": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "integer",
						"format": "int32",
						"nullable": true
					},
					"detail": {
						"type": "string",
						"nullable": true
					},
					"instance": {
						"type": "string",
						"nullable": true
					}
				}
			},
...
			"SimpleDeploymentDto": {
				"required": [
					"createdAt",
					"updatedAt",
					"id",
					"state",
					"stackId",
					"stackName",
					"projectId",
					"environmentId"
				],
				"type": "object",
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"state": {
						"$ref": "#/components/schemas/DeploymentStatusDto"
					},
					"stackId": {
						"type": "string",
						"format": "uuid"
					},
					"stackName": {
						"type": "string"
					},
					"projectId": {
						"type": "string",
						"format": "uuid"
					},
					"environmentId": {
						"type": "string",
						"format": "uuid"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				}
			},
...
			"StackCreatedDto": {
				"required": ["id"],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					}
				}
			},
			"StackDetailDto": {
				"required": [
					"id",
					"name",
					"description",
					"source",
					"destination",
					"health",
					"services"
				],
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"services": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/StackServiceDto"
						}
					},
					"source": {
						"$ref": "#/components/schemas/StackSourceDto"
					},
					"destination": {
						"$ref": "#/components/schemas/StackServerDto"
					},
					"health": {
						"$ref": "#/components/schemas/EntityHealthDto"
					}
				}
			},
...
			"UpdateStackCommand": {
				"required": ["stackId"],
				"type": "object",
				"properties": {
					"stackId": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string",
						"default": null,
						"nullable": true
					},
					"description": {
						"type": "string",
						"default": null,
						"nullable": true
					},
					"gitPublicSettings": {
						"$ref": "#/components/schemas/UpdateStackGitPublicSettings"
					},
					"githubSettings": {
						"$ref": "#/components/schemas/UpdateStackGithubSettings"
					}
				}
			},
...
		}
	},
...
}